Merged PR 4974: Update npm run dev
This PR updates `npm run dev` to work the same way `npm run dev-hashed` does, with the exception of it removes the hash so that while working in dev mode you don't have to stop and restart the app.

This also fixes a minor bug regarding pipes being vertically misaligned (noted in Slack by AK, no bug ticket was assigned) as a test to ensure the app refreshed properly without the need to restart.