Merged PR 4995: Improve Static Gen to include more hooks.
As we move onto sections that are outside of the main article, I'm in need of more hooks for my static generator. IE: Alerts, Breadcrumbs, etc. I've add a main before and after hook for the static html generator.

This update also keeps `npm run dev` and `npm run dev-hashed` in sync so that when changes are made it effects both at the same time.

*There may be potential merge conflicts with existing branches that pulled off of dev before this update.  Please tag me and I'll resolve them if they pop up.*